Fix windows compilation issues.
