+++ crates/warp_terminal/src/focus_env.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 use std::ffi::OsString;
 
-use crate::channel::ChannelState;
+use warp_core::channel::ChannelState;
 
 pub(crate) const FOCUS_URL_ENV: &str = "WARP_FOCUS_URL";
 pub(crate) const TERMINAL_SESSION_UUID_ENV: &str = "WARP_TERMINAL_SESSION_UUID";
@@ -13,10 +13,7 @@ pub(crate) fn session_focus_url(session_uuid_hex: &str) -> String {
     )
 }
 
-pub(crate) fn add_session_focus_env_vars(
-    env_vars: &mut HashMap<OsString, OsString>,
-    session_uuid: &[u8],
-) {
+pub fn add_session_focus_env_vars(env_vars: &mut HashMap<OsString, OsString>, session_uuid: &[u8]) {
     let session_uuid_hex = hex::encode(session_uuid);
     env_vars.insert(
         OsString::from(TERMINAL_SESSION_UUID_ENV),
