+++ crates/warp_terminal/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bootstrap;
 pub mod event;
 pub mod event_listener;
+pub mod focus_env;
 #[cfg(not(target_family = "wasm"))]
 pub mod local_tty;
 pub mod model;
