+++ crates/warp_terminal/src/local_tty/shell.rs
@@ -9,6 +9,8 @@ use typed_path::UnixPathBuf;
 use warp_core::channel::{Channel, ChannelState};
 use warp_core::session_id::SessionId;
 use warp_errors::report_error;
+#[cfg(windows)]
+use warp_util::path::windows::{powershell_5_path, powershell_7_path, wsl_path};
 use warp_util::path::{
     canonicalize_git_bash_path, is_msys2_path, resolve_executable, warp_shell_path,
 };
@@ -16,8 +18,6 @@ use warp_util::path::{
 use crate::bootstrap::{generate_session_id, init_shell_script_for_shell};
 use crate::local_tty::docker_sandbox::DockerSandboxShellStarter;
 use crate::shell::{ShellLaunchData, ShellName, ShellType};
-#[cfg(windows)]
-use crate::util::windows::{powershell_5_path, powershell_7_path, wsl_path};
 
 pub const ZSH_SHELL_PATH: &str = "/bin/zsh";
 pub const BASH_SHELL_PATH: &str = "/bin/bash";
