+++ crates/warp_terminal/src/local_tty/windows/environment.rs
@@ -4,19 +4,20 @@ use std::os::windows::ffi::{OsStrExt, OsStringExt};
 
 use itertools::Itertools;
 use warp_core::channel::ChannelState;
-use warp_core::cli_agent_protocol::{WARP_CLI_AGENT_PROTOCOL_VERSION_ENV, WARP_CLIENT_VERSION_ENV};
+use warp_core::cli_agent_protocol::{
+    CLI_AGENT_PROTOCOL_VERSION, WARP_CLI_AGENT_PROTOCOL_VERSION_ENV, WARP_CLIENT_VERSION_ENV,
+};
 use warp_core::features::FeatureFlag;
+use warp_core::safe_info;
 use windows::Win32::System::Environment::ExpandEnvironmentStringsW;
 use windows::core::{HSTRING, PCWSTR};
 use winreg::enums::{HKEY_CURRENT_USER, HKEY_LOCAL_MACHINE, RegType};
 use winreg::types::FromRegValue;
 use winreg::{RegKey, RegValue};
 
-use crate::cli_agent_sessions::event::current_protocol_version;
 use crate::focus_env::{FOCUS_URL_ENV, TERMINAL_SESSION_UUID_ENV};
 use crate::local_tty::PtyOptions;
 use crate::local_tty::shell::{ShellStarter, extra_path_entries, ssh_socket_dir};
-use crate::safe_info;
 
 const HONOR_PS1_NAME: &str = "WARP_HONOR_PS1";
 const PROMPT_NODE_VERSION_ENABLED_NAME: &str = "WARP_PROMPT_NODE_VERSION_ENABLED";
@@ -139,7 +140,7 @@ pub(super) fn get_shell_environment_variables(options: &PtyOptions) -> Vec<u16> 
             map_key(WARP_CLI_AGENT_PROTOCOL_VERSION_ENV.into()),
             EnvEntry {
                 preferred_key: WARP_CLI_AGENT_PROTOCOL_VERSION_ENV.into(),
-                value: current_protocol_version().to_string().into(),
+                value: CLI_AGENT_PROTOCOL_VERSION.to_string().into(),
             },
         );
     }
